New stable version that going to be tagged for package tests.
Fix problem with open file handles. The forked process was keeping the
file handle of the OPC UA port. If the watchdog server was killed still
the started processes were blocking the watchdog server port. Now all
file handles except stdin stdout and stderr will be closed when exec is
called. 

Prepare systemd based automatic start of the watchdog. That is not yet
completed. 

Move the calling of the xml genereator to the debug builds. When using
pbuilder it has no access to /sys/devices... and calling the xml
genereator will crash.