Don't build a library and remove boost process option.
Since the package will be build as bin package no library will be build.
If one wants to use some functionality from here it should be put in a
different project - like with the LLRF library and the server.
From now I removed the boost process option, since I don't maintain it
anyway and I'm not sure if all current features can be realised using
boost process. Could be an option for the future when boost>1.64 is
available in the OS.

Minor changes that came up during deb file building.
- Fix some wrong names and systemd settings.
- Writing a PID file is not needed by systemd.
- Update documentation.
- Adopt conf files and remove files not needed.
- Put config files to /etc/ and leave service in /var/lib
- Add explaination why files are put to /etc/