Keep processes running when the Watchdog is stopped.
When the Watchdog process is killed systemd was also stopping all
processes started by the Watchdog. This is now avoided. When the
Watchdog is stopped all running processes are stopped. This is changed
too, since the Watchdog is stopped every time a new version is
installed. Now in any case the processes started by the watchdog are
kept alive no matter what happens to the Watchdog.