Write PID files to /tmp. Fix counter. Fix and extend tests.
Now the PID files are written to /tmp. This is because by doing so it is
guaranteed that they are removed after remoot. Else it could happen that
a PID is read from the file and after reboot an arbitrary process was
started with that PID. Thus the watchdog assumes the process to be
started is already running, which is not true since a different process
uses the PID read from the PID file.