Add test for restart counter.
It turned it that the test so far did only check if the running state is
correct after max restarts is reached, if the process terminates before
the next trigger after reaching max restarts. But usually the process
terminates much later. In this case the running state is not correct so
far.