BugFix: Logfile data is not updated when the log file is deleted.
Once the log file is deleted and the process is started again a new log
file is created. Here still the last information of the old file was
shown, since the file buffer did recognize the file deletion. By
reopening the filebuffer everytime it is ensured to always read the
current log file.